requirements.txt
streamlit==1.29.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.0
joblib==1.3.2